requirements.txt
asgiref==3.5.0
certifi==2021.10.8
charset-normalizer==2.0.12
defusedxml==0.7.1
diff-match-patch==20200713
dj-database-url==0.5.0
dj-static==0.0.6
Django==3.2.2
django-import-export==2.8.0
django-widget-tweaks==1.4.12
et-xmlfile==1.1.0
gunicorn==20.1.0
heroku==0.1.4
idna==3.3
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.0.10
Pillow==9.1.0
psycopg2==2.9.3
psycopg2-binary==2.9.3
python-dateutil==1.5
python-decouple==3.6
pytz==2022.1
PyYAML==6.0
requests==2.27.1
sqlparse==0.4.2
static3==0.7.0
tablib==3.2.1
tzdata==2022.1
urllib3==1.26.9
whitenoise==6.0.0
xlrd==2.0.1
xlwt==1.3.0
